+++ src/MFO.CatalogService.Infrastructure/Repositories/CategoryRepository.cs
@@ -17,7 +17,7 @@ public class CategoryRepository : ICategoryRepository
     public async Task<Category?> GetCategoryByIdAsync(Guid id, CancellationToken cancellationToken)
         => await _db.Categories.FindAsync([id], cancellationToken);
 
-    public async Task<List<Category>> GetAllCategoriesAsync(CancellationToken cancellationToken)
+    public async Task<IReadOnlyList<Category>> GetAllCategoriesAsync(CancellationToken cancellationToken)
         => await _db.Categories.ToListAsync(cancellationToken);
 
     public async Task<Category> AddCategoryAsync(Category category, CancellationToken cancellationToken)
@@ -34,8 +34,14 @@ public class CategoryRepository : ICategoryRepository
         return category;
     }
 
-    public async Task<bool> DeleteCategoryAsync(Category category, CancellationToken cancellationToken)
+    public async Task<bool> DeleteCategoryAsync(Guid categoryId, CancellationToken cancellationToken)
     {
+        var category = await _db.Categories.FindAsync([categoryId], cancellationToken);
+        if (category is null)
+        {
+            return false;
+        }
+
         _db.Categories.Remove(category);
         await _db.SaveChangesAsync(cancellationToken);
         return true;
